requirements.txt
appdirs==1.4.3
attrdict==2.0.0
backcall==0.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
colorama==0.3.9
cytoolz==0.9.0.1
decorator==4.3.0
eth-abi==1.1.1
eth-account==0.2.3
eth-hash==0.1.4
eth-keyfile==0.5.1
eth-keys==0.2.0b3
eth-rlp==0.1.2
eth-utils==1.0.3
hexbytes==0.1.0
idna==2.7
ipfsapi==0.4.3
ipython==6.5.0
ipython-genutils==0.2.0
jedi==0.12.1
lru-dict==1.1.6
parsimonious==0.8.0
parso==0.3.1
pickleshare==0.7.4
prompt-toolkit==2.0.3
py-solc==3.1.0
pycryptodome==3.6.4
Pygments==2.2.0
pypiwin32==223
pywin32==223
requests==2.19.1
rlp==1.0.1
semantic-version==2.6.0
simplegeneric==0.8.1
six==1.11.0
toolz==0.9.0
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
web3==4.4.1
websockets==5.0.1
